Presenter should not be mocked. When testing the presenter, I'm also really testing the view as well. So my tests here cover MoviesView's methods as well as presenter's methods. In setup() I now initialize a new executor and presenter, as well as go ahead and attach the mocked view to the presenter.
shouldGetMovies() creates a dummy arraylist with dummy data that is then returned by mocked MDP's getMovieData(). Presenter presents, and I verify that the view never shows an error and then that it shows movies as set.

shouldShowErrorTest() throws an exception when setMovies(empty arraylist) is called on the mocked view. Verifies movies are never set/shown and that it does show error.